requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.17.0
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
iso8601==1.1.0
magic-filter==1.0.12
multidict==6.0.5
pydantic==1.10.14
pypika-tortoise==0.1.6
pytz==2024.1
tortoise-orm==0.19.2
typing_extensions==4.10.0
yarl==1.9.4
firemelon==1.1.4